moved geo-map folder into features
