+++ src/features/geo-map/lib/calculate-bounds-from-points.ts
