+++ src/features/geo-map/lib/is-valid-point.ts
