+++ src/features/visualization-layouts/visualization.tsx
@@ -4,9 +4,9 @@ import { useAppSelector } from '@/app/store';
 import { selectEntities, selectEvents } from '@/app/store/intavia.slice';
 import type { Visualization } from '@/features/common/visualization.slice';
 import { NetworkComponent } from '@/features/ego-network/network-component';
+import { GeoMapWrapper } from '@/features/geo-map/geo-map-wrapper';
 //import { StoryTimeline } from '@/features/storycreator/story-timeline';
 import { TimelineComponent } from '@/features/timelineV2/timelineComponent';
-import { GeoMapWrapper } from '@/features/visualizations/geo-map/geo-map-wrapper';
 import { useElementDimensions } from '@/lib/use-element-dimensions';
 import { useElementRef } from '@/lib/use-element-ref';
